+++ howard/objects/database.py
@@ -17,6 +17,7 @@ from tempfile import TemporaryDirectory
 from typing import Optional, Union
 
 from howard.functions.commons import (
+    cast_columns_query,
     load_duckdb_extension,
     get_file_format,
     full_path,
@@ -2558,39 +2559,8 @@ class Database:
                         {order_by_sql}
                     """
 
-                # Find structure, with DESCRIBE with query, of query with columns and type like VARCHAR or VARCHAR[], or FLOAT or FLOAT[]
-                # and create a dictionary with column name and type
-                query_sql_structure = f"""
-                    DESCRIBE ({query})
-                """
-                result_describe = self.conn.execute(query_sql_structure)
-                description_dict = {
-                    row[0]: {"type": row[1]} for row in result_describe.fetchall()
-                }
-                # log.debug(f"description_dict={description_dict}")
-
-                # Add a select to aggregate list to string if needed
-                list_columns = []
-                for column in description_dict:
-                    # column_type = description_dict[column]["type"]
-                    column_type = description_dict.get(column, {}).get(
-                        "type", "VARCHAR"
-                    )
-                    if column_type.endswith("[]"):
-                        column_type = "VARCHAR"
-                        list_columns.append(
-                            f""" list_aggregate("{column}", 'string_agg', ',') AS '{column}' """
-                        )
-                    else:
-                        list_columns.append(f""" "{column}" """)
-
-                # RE-Query
-                # log.debug(f"Export query 1: {query}")
-                query = f"""
-                    SELECT {",".join(list_columns)}
-                    FROM ({query})
-                """
-                # log.debug(f"Export query 2: {query}")
+                # Cast query column to be able to export into a flat file
+                query = cast_columns_query(query=query, conn=self.conn)
 
                 # Test empty query
                 df = self.conn.execute(query).fetch_record_batch(1)